Skywalker now outputs floating point values in full precision.
This commit also changes sw_output_set_array so the size parameter
is a size_t and not a pointer. This required adding a small Fortran
wrapper and updating the API documentation.

We don't expect this change to be disruptive, because the C API is
mostly being used to support the Fortran and C++ APIs at the moment.